+++ CHANGELOG.md
@@ -13,6 +13,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - PSR-11 implementation in Application
 - BootableProviderInterface
 
+### Changed
+
+- Run calls always boot
+- if no uSilex.responseEmitter, the body of the response is emitted
 
 ### Fixed